requirement.txt
streamlit
ollama
Pillow

